added log times, script ran in 19 min
